orderSync - confirm first on
WooCommerce webhook sends data to receiver code on Myxfin application
Format the Json data into what it will look like on the real database
Is confirm-first on? (we still have to find a way to store this value) If yes,
Save Json into the woocommerce landing table. This is a landing table that will save the json data